[UPDATE]: security in journal controller & service

+++ src/main/java/com/digest/dertyu/Controllers/JournalEntryControllerV2.java
@@ -9,6 +9,9 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
@@ -20,10 +23,13 @@ public class JournalEntryControllerV2 {
     //private Map<Long,JournalEntity>journalEntityMap=new HashMap<>();
     private final JournalEntryService journalEntryService;
     private final UserEntryService userEntryService;
-    @GetMapping("/get/{name}")
-    public ResponseEntity<List<JournalEntity>>getAllJournalEntriesOfUser(@PathVariable String name){
-        User byUserName = userEntryService.findByUserName(name);
-        List<JournalEntity> journalEntities = byUserName.getJournalEntities();
+    @GetMapping("/get")
+    public ResponseEntity<List<JournalEntity>>getAllJournalEntriesOfUser(){
+        //User byUserName = userEntryService.findByUserName(name);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
+        User user=userEntryService.findByUserName(name);
+        List<JournalEntity> journalEntities = user.getJournalEntities();
         try {
             return new ResponseEntity<>(journalEntities,HttpStatus.OK);
         } catch (Exception e) {
@@ -31,9 +37,10 @@ public class JournalEntryControllerV2 {
         }
     }
 
-    @PostMapping("/post/{name}")
-    public ResponseEntity<JournalEntity>createJournal(@RequestBody JournalEntity journalEntity,
-                                                      @PathVariable String name){
+    @PostMapping("/post")
+    public ResponseEntity<JournalEntity>createJournal(@RequestBody JournalEntity journalEntity){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
         try {
             return new ResponseEntity<>(journalEntryService.createEntry(journalEntity,name),HttpStatus.CREATED);
         } catch (Exception e) {
@@ -44,29 +51,50 @@ public class JournalEntryControllerV2 {
     @GetMapping("id/{myId}")
     public ResponseEntity<Optional<JournalEntity>>getbyId(@PathVariable ObjectId myId){
        // return journalEntryService.getByJournalId(myId).orElse(null);
-        Optional<JournalEntity> byJournalId = journalEntryService.getByJournalId(myId);
-        if(byJournalId.isPresent()){
-            return new ResponseEntity<>(byJournalId, HttpStatus.FOUND);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
+        User user=userEntryService.findByUserName(name);
+        List<JournalEntity> collect = user.getJournalEntities().stream().filter(x -> x.getId().equals(myId)).toList();
+        if(!collect.isEmpty()){
+            Optional<JournalEntity> byJournalId = journalEntryService.getByJournalId(myId);
+            if(byJournalId.isPresent()){
+                return new ResponseEntity<>(byJournalId, HttpStatus.FOUND);
+            }
         }
         return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @DeleteMapping("/{id}/{name}")
-    public ResponseEntity<Boolean>deleteById(@PathVariable ObjectId id,@PathVariable String name){
-        try {
-            return new ResponseEntity<>(journalEntryService.deleteByIdherE(id,name),HttpStatus.FOUND);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Boolean>deleteById(@PathVariable ObjectId id){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
+        User user=userEntryService.findByUserName(name);
+        List<JournalEntity> collect = user.getJournalEntities().stream().filter(x -> x.getId().
+                equals(id)).toList();
+        if(!collect.isEmpty()){
+            try {
+                return new ResponseEntity<>(journalEntryService.deleteByIdherE(id,name),HttpStatus.FOUND);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-    @PatchMapping("/update/{id}/{name}")
+    @PatchMapping("/update/{id}")
     public ResponseEntity<JournalEntity>updateBYId(@PathVariable ObjectId id,
-                                                   @RequestBody JournalEntity journalEntity,
-                                                   @PathVariable String name){
-        try {
-            return new ResponseEntity<>(journalEntryService.updateyhis(id,journalEntity,name),HttpStatus.ACCEPTED);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+                                                   @RequestBody JournalEntity journalEntity){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String name = authentication.getName();
+        User byUserName = userEntryService.findByUserName(name);
+        List<JournalEntity> collect = byUserName.getJournalEntities().stream().filter(x ->
+                        x.getId().equals(id)).toList();
+        if(!collect.isEmpty()){
+            try {
+                return new ResponseEntity<>(journalEntryService.updateyhis(id,journalEntity,name),HttpStatus.ACCEPTED);
+            } catch (Exception e) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
